Kill the monkey! (#322)
