+++ lib/ddtrace/contrib/active_record/patcher.rb
@@ -2,7 +2,6 @@ module Datadog
   module Contrib
     module ActiveRecord
       # Patcher enables patching of 'active_record' module.
-      # This is used in monkey.rb to manually apply patches
       module Patcher
         include Base
         register_as :active_record, auto_patch: false
