+++ lib/ddtrace/contrib/elasticsearch/patcher.rb
@@ -11,7 +11,6 @@ module Datadog
       SERVICE = 'elasticsearch'.freeze
 
       # Patcher enables patching of 'elasticsearch/transport' module.
-      # This is used in monkey.rb to automatically apply patches
       module Patcher
         include Base
         register_as :elasticsearch, auto_patch: true
@@ -28,7 +27,6 @@ module Datadog
             begin
               require 'uri'
               require 'json'
-              require 'ddtrace/monkey'
               require 'ddtrace/pin'
               require 'ddtrace/ext/app_types'
               require 'ddtrace/contrib/elasticsearch/quantize'
@@ -49,7 +47,7 @@ module Datadog
           # rubocop:disable Metrics/BlockLength
           ::Elasticsearch::Transport::Client.class_eval do
             alias_method :initialize_without_datadog, :initialize
-            Datadog::Monkey.without_warnings do
+            Datadog::Patcher.without_warnings do
               remove_method :initialize
             end
