+++ test/contrib/elasticsearch/test_helper.rb
@@ -3,8 +3,6 @@ require 'net/http'
 require 'ddtrace'
 require 'elasticsearch/transport'
 
-Datadog::Monkey.patch_module(:elasticsearch)
-
 def wait_http_server(server, delay)
   delay.times do
     uri = URI(server + '/')
