+++ test/contrib/http/integration_test.rb
@@ -12,6 +12,10 @@ class HTTPIntegrationTest < Minitest::Test
   def setup
     skip unless ENV['TEST_DATADOG_INTEGRATION'] # requires a running agent
 
+    Datadog.configure do |c|
+      c.use :http
+    end
+
     # Here we use the default tracer (to make a real integration test)
     @tracer = Datadog.tracer
