+++ test/contrib/mongodb/test_helper.rb
@@ -1,4 +1,2 @@
 require 'ddtrace'
 require 'mongo'
-
-Datadog::Monkey.patch_module(:mongo)
